Update Dockerfile to use rstudio 2021.09.0+351
Nice work on this Dockerfile recipe. It still works with the latest RStudio and a minor tweak to the dependency list, so I thought I'd pass it along in case you want to update them.

The only problem I've had is that (naturally) RStudio can't launch browser tabs when it's running inside a container. I found this idea https://stackoverflow.com/a/63719458/6058999 about using a named pipe for that, so I'm working on a docker-compose/makefile wrapper to set that up